Print the smallest floating point representation.
This commmit makes a change where floating points are now
printed as their smallest textual representation. Before, all
numbers were always ddd.fffff. But now it may be printed with an
exponent, if and when that will make it smaller. This affects wkt
and geojson outputs. The reason for this change is to save space,
both network and disk, but without losing any accuracy.

For example, the number 5000000 will be now be 5e6.

To disable this feature call tg_env_set_print_fixed_floats(true);